+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 
+- [#359](https://github.com/clojure-emacs/orchard/pull/359): Print: protect printing from broken eductions.
+
 ## 0.37.0 (2025-09-19)
 
 - [#353](https://github.com/clojure-emacs/orchard/pull/353): Stacktrace: flag Clojure functions as duplicate.